+++ app/src/androidTest/java/com/example/campusguide/DirectionsExploratoryTest.kt
@@ -2,10 +2,10 @@ package com.example.campusguide
 
 import androidx.test.internal.runner.junit4.AndroidJUnit4ClassRunner
 import androidx.test.platform.app.InstrumentationRegistry.getInstrumentation
+import com.example.campusguide.directions.Directions
 import kotlinx.coroutines.runBlocking
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
-import org.hamcrest.Matchers.greaterThan
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -16,7 +16,8 @@ class DirectionsExploratoryTest {
         runBlocking {
             val context = getInstrumentation().targetContext
 
-            val response = Directions(context).getDirections("Hall Building", "Faubourg Building")
+            val response = Directions(context)
+                .getDirections("Concordia Hall Building", "Faubourg Building")
             val eq = response?.status.equals("OK")
             assertThat(eq, equalTo(true))
         }
