+++ app/src/main/java/com/example/campusguide/directions/Directions.kt
@@ -1,12 +1,12 @@
-package com.example.campusguide
+package com.example.campusguide.directions
 
 import android.content.Context
 import com.android.volley.Request
 import com.android.volley.Response
 import com.android.volley.toolbox.JsonObjectRequest
 import com.beust.klaxon.Klaxon
+import com.example.campusguide.R
 import com.example.campusguide.utils.RequestDispatcher
-import com.example.campusguide.utils.dtos.GoogleDirectionsAPIResponse
 import org.json.JSONObject
 import java.net.URLEncoder
 import kotlin.coroutines.resume
@@ -17,8 +17,16 @@ class Directions {
     constructor(applicationContext: Context) {
         context = applicationContext
     }
+
+    /**
+     * Gets directions from the Google API
+     */
     suspend fun getDirections(startLocation: String, endLocation: String) = suspendCoroutine<GoogleDirectionsAPIResponse?> { cont ->
         val apiKey = context.resources.getString(R.string.google_maps_key)
+
+        /**
+         * Start and end will be placed in query params, so they must be urlEncoded.
+         */
         val startEncoded = URLEncoder.encode(startLocation, "UTF-8")
         val endEncoded = URLEncoder.encode(endLocation, "UTF-8")
 
@@ -26,7 +34,6 @@ class Directions {
 
         val request = JsonObjectRequest(Request.Method.GET, url, null,
             Response.Listener<JSONObject> { response ->
-
                 val responseObj = Klaxon().parse<GoogleDirectionsAPIResponse>(response.toString(0))
                 cont.resume(responseObj)
             },
